mcp-context/requirements.txt
mcp>=1.11.0
pydantic>=2.5.0
asyncpg>=0.29.0
sqlalchemy[asyncio]>=2.0.25
alembic>=1.13.0
redis[hiredis]>=5.0.0
python-jose[cryptography]>=3.3.0
python-dotenv>=1.0.0
structlog>=24.1.0
tenacity>=8.2.0
prometheus-client>=0.19.0
python-multipart>=0.0.6
httpx>=0.27.0
tiktoken>=0.5.0